fibonacci betty error and more

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,8 @@
 int main(void)
 {
 	int a, b, d;
-	long int c;
+
+	int c;
 	a = 1;
 	b = 2;
 	c = 0;
@@ -16,7 +17,7 @@ int main(void)
 	while (d <= 48)
 	{
 		c = a + b;
-		printf("%li", c);
+		printf("%i", c);
 		if (d != 48)
 			printf(", ");
 		a = b;
